separate common tests from tests of optional interface methods
