+++ tests/test_memcached_cache.py
@@ -1,6 +1,8 @@
 import pytest
-
+from conftest import ClearTests
 from conftest import CommonTests
+from conftest import HasTests
+
 from cachelib import MemcachedCache
 
 
@@ -15,16 +17,5 @@ def cache_factory(request):
 
 
 @pytest.mark.usefixtures("memcached_server")
-class TestMemcachedCache(CommonTests):
-    def test_has(self):
-        cache = self.cache_factory()
-        assert cache.set_many(self.sample_pairs)
-        for k in self.sample_pairs:
-            assert cache.has(k)
-        assert not cache.has("unknown")
-
-    def test_clear(self):
-        cache = self.cache_factory()
-        assert cache.set_many(self.sample_pairs)
-        assert cache.clear()
-        assert not any(cache.get_many(*self.sample_pairs))
+class TestMemcachedCache(CommonTests, ClearTests, HasTests):
+    pass
